requirements.txt
azure-ai-formrecognizer==3.0.0
azure-ai-textanalytics==5.1.0b4
azure-cognitiveservices-vision-computervision==0.7.0
azure-cognitiveservices-vision-contentmoderator==1.0.0
azure-cognitiveservices-vision-customvision==3.1.0
azure-common==1.1.26
azure-core==1.10.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
isodate==0.6.0
joblib==1.0.0
msrest==0.6.21
numpy==1.19.5
oauthlib==3.1.0
pandas==1.2.1
Pillow==8.1.0
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
scikit-learn==0.24.1
scipy==1.6.0
sidetable==0.8.0
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
urllib3==1.26.3
